refactor: user package
